% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/score_indicators.R
\name{aucroc}
\alias{aucroc}
\title{Calculate Area Under ROC Curve}
\usage{
aucroc(score, target)
}
\arguments{
\item{score}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}
}
\value{
The AUROC Curve
}
\description{
Calculate Area Under ROC Curve
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

aucroc(score, target)
}

