% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autobinning.R
\name{autobinning}
\alias{autobinning}
\title{Autobinning}
\usage{
autobinning(target, variable)
}
\arguments{
\item{target}{A numeric binary vector {0,1}}

\item{variable}{A variable}
}
\value{
A list of elements
}
\description{
Automatic supervised binning
}
\examples{
data("credit")

target <- credit$bad
variable <- credit$age

autobinning(credit$bad, credit$age)
}

