% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binning.R
\name{bin_sup}
\alias{bin_sup}
\title{Supervised Binning}
\usage{
bin_sup(variable, target, min.p = 0.5)
}
\arguments{
\item{variable}{A variable}

\item{target}{A numeric binary vector {0,1}}

\item{min.p}{Minimal proportion in a group}
}
\value{
A list of elements
}
\description{
Automatic supervised binning
}
\examples{
data("credit")

variable <-  credit$age
target <- credit$bad

bin_sup(variable, target)

variable <- credit$marital_status

bin_sup(variable, target)
}

