% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binning.R
\docType{data}
\name{Binning}
\alias{Binning}
\title{Binning Class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x061d1220> 
 - attr(*, "name")= chr "Binning_generator"
}}
\usage{
Binning
}
\description{
Binning class.
}
\keyword{datasets}

