% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conf_matrix.R
\name{conf_matrix}
\alias{conf_matrix}
\title{Create confusion matrix}
\usage{
conf_matrix(pred_class, target)
}
\arguments{
\item{pred_class}{A vector numeric binary (0, 1) of predictions}

\item{target}{A numeric binary vector (0, 1)}
}
\value{
A list
}
\description{
Create confusion matrix
}
\examples{
conf_matrix()
}
\references{
http://www2.cs.uregina.ca/~dbd/cs831/notes/confusion_matrix/confusion_matrix.html
}

