% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{divergence}
\alias{divergence}
\title{Calculate Divergence}
\usage{
divergence(target, score)
}
\arguments{
\item{target}{A numeric binary vector (0, 1)}

\item{score}{A numeric vector containing scores or probabilities}
}
\value{
The Gini Coefficient
}
\description{
Calculate the divergence between empirical distributions
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

divergence(target, score)
}

