% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_summ.R
\name{ez_summ}
\alias{ez_summ}
\title{Generate Summary Tables for data frames}
\usage{
ez_summ(df, target_name = NULL, verbose = TRUE, ...)
}
\arguments{
\item{df}{A data frame.}

\item{target_name}{Target name if bivariate summary is required}

\item{...}{A list of additional arguments}
}
\value{
A list
}
\description{
Generate Summary Tables for data frames
}
\examples{

data("credit")

credit \%>\% ez_summ()
credit \%>\% ez_summ(target_name = "bad")



\dontrun {

df <- credit

set.seed(123)

for(var in names(df)){
  message(var)
  df[[var]] <- ifelse(runif(nrow(df)) < 0.05, NA, df[[var]])
}

df <- df \%>\% filter(!is.na(bad))
df \%>\% ez_summ()
df \%>\% ez_summ_biv(target_name = "bad")
df \%>\% ez_summ(target_name = "bad")
}


}

