% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_summ.R
\name{ez_summ_cat}
\alias{ez_summ_cat}
\title{Generate Tables for categorical variables}
\usage{
ez_summ_cat(df)
}
\arguments{
\item{df}{A data frame.}
}
\description{
Generate Tables for categorical variables
}
\examples{

data("credit")

credit \%>\% select_categorical() \%>\% ez_summ_cat()

}

