% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_summ.R
\name{ez_summ_num}
\alias{ez_summ_num}
\title{Generate Summary Tables for numercial variables}
\usage{
ez_summ_num(df, na.rm = TRUE, probs = c(0.01, 0.05, 1:9/10, 0.95, 0.99))
}
\arguments{
\item{df}{A data frame.}

\item{na.rm}{A logical indicating whether missing values should be removed.}

\item{probs}{numeric vector of probabilities with values in [0,1].}
}
\description{
Generate Summary Tables for numercial variables
}
\examples{

data("credit")

credit \%>\% select_numeric() \%>\% ez_summ_num()

}

