% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/riskr.R
\name{fight}
\alias{fight}
\title{Simulate a risk campaign}
\usage{
fight(a, d, sims = 1000)
}
\arguments{
\item{a}{The number of attacking units (remember to minus one to be left behind).}

\item{d}{A vector giving the number of defending units in successive defenders countries.}
}
\description{
Given a succession of defending countries, what are your odds of winning and how far will you get?
}

