% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{ft}
\alias{ft}
\title{Frecueny table from a vector}
\usage{
ft(x, order.by.count = TRUE)
}
\arguments{
\item{x}{A vector}

\item{order.by.count}{Logical value if the df response have to be order by count.}
}
\value{
The frecuency table of \code{x}. Frecuencies and percents are returned in a
\code{data.frame} object (adding \code{tbl_df} class).
}
\description{
Frecueny table from a vector
}
\examples{
set.seed(1313)
x <- sample(LETTERS[1:10], size = 100, prob = 1:10/10, replace = TRUE)
ft(x)
}

