% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/performance.R
\name{gain}
\alias{gain}
\title{Calculate Gains}
\usage{
gain(score, target, percents = c(0.1, 0.2, 0.3, 0.4, 0.5))
}
\arguments{
\item{score}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}

\item{percents}{Values to calculate the gain}
}
\value{
The Gini Coefficient
}
\description{
Calculate Gains
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

gain(score, target)
}

