% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_analysis.R
\name{gg_ba2}
\alias{gg_ba2}
\title{Plot Bivariate Analysis (2)}
\usage{
gg_ba2(variable, target, labels = TRUE, order.by = NULL)
}
\arguments{
\item{variable}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}

\item{labels}{A par}

\item{order.by}{A par}
}
\value{
A ggplot2 object
}
\description{
A minimal version for \emph{gg_ba}
}
\examples{
data("credit")

variable <- as.character(credit$marital_status)
target <- credit$bad

gg_ba2(variable, target)
gg_ba2(variable, target, labels = FALSE)
gg_ba2(variable, target, order.by = "odds")

}

