% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gg_plots_perf.R
\name{gg_lift}
\alias{gg_lift}
\title{Plot Lift Chart}
\usage{
gg_lift(target, score)
}
\arguments{
\item{target}{A numeric binary vector (0, 1)}

\item{score}{A numeric vector containing scores or probabilities}
}
\value{
The plot
}
\description{
Return a plot with the lift chart
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

gg_lift(target, score)
}
\references{
http://www2.cs.uregina.ca/~dbd/cs831/notes/lift_chart/lift_chart.html http://www.saedsayad.com/model_evaluation_c.htm
}

