% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gg_plots_perf.R
\name{gg_perf}
\alias{gg_perf}
\title{Plot Performance}
\usage{
gg_perf(target, score)
}
\arguments{
\item{target}{A numeric binary vector (0, 1)}

\item{score}{A numeric vector containing scores or probabilities}
}
\value{
The plot
}
\description{
Return a plot with gg_roc, gg_dists, gg_cum, gg_gain
}
\examples{
data(predictions)

score <- 1000 * predictions$score
target <- predictions$target

gg_perf(target, score)
}

