% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{gg_roc}
\alias{gg_roc}
\title{Plot AUC Curve}
\usage{
gg_roc(score, target)
}
\arguments{
\item{score}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}
}
\value{
the plot
}
\description{
Plot AUC Curve
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

gg_roc(score, target)
}

