% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_analysis.R
\name{ggs_biv}
\alias{ggs_biv}
\title{Plot bivariates gg_ba}
\usage{
ggs_biv(df, target_name, verbose = TRUE)
}
\description{
Plot bivariates gg_ba
}
\examples{

data("credit")

library("ggplot2")
library("ggthemes")
theme_set(theme_fivethirtyeight(base_size = 11) +
theme(rect = element_rect(fill = "white"),
axis.title = element_text(colour = "grey30"),
axis.title.y = element_text(angle = 90),
strip.background = element_rect(fill = "#434348"),
strip.text = element_text(color = "#F0F0F0"),
plot.title = element_text(face = "plain", size = structure(1.2, class = "rel")),
panel.margin.x =  grid::unit(1, "cm"),
panel.margin.y =  grid::unit(1, "cm")))
update_geom_defaults("line", list(colour = "#434348", size = 1.05))
update_geom_defaults("point", list(colour = "#434348", size = 3))
update_geom_defaults("bar", list(fill = "#7cb5ec"))
update_geom_defaults("text", list(size = 4, colour = "gray30"))

plots <- ggs_biv(credit, "bad")

bquiet = lapply(plots, print)
dev.off()

}

