% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/score_indicators.R
\name{gini}
\alias{gini}
\title{Calculate Gini Coefficient}
\usage{
gini(score, target)
}
\arguments{
\item{score}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}
}
\value{
The Gini Coefficient
}
\description{
Calculate Gini Coefficient
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

gini(score, target)
}

