% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/score_indicators.R
\name{ks}
\alias{ks}
\title{Calculate Kolmogorv}
\usage{
ks(score, target)
}
\arguments{
\item{score}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}
}
\value{
The KS statistic
}
\description{
Calculate Kolmogorv
}
\examples{
ks(score, target)
}

