% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/performance.R
\name{ks2}
\alias{ks2}
\title{Calculate Kolmogorov-Smirnov statistic}
\usage{
ks2(score1, score2)
}
\arguments{
\item{score1}{A numeric vector containing scores or probabilities}

\item{score2}{A numeric vector containing scores or probabilities}
}
\value{
The KS statistic
}
\description{
Calculate Kolmogorov-Smirnov statistic
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

score1 <- score[target == 1]
score1 <- score[target != 1]

ks2(score, target)
}

