% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linear_formula.R
\name{linear_formula}
\alias{linear_formula}
\title{Write the Linear Combination}
\usage{
linear_formula(model)
}
\arguments{
\item{model}{A model}
}
\value{
A string
}
\description{
Write the linear combination
}
\examples{
model <- glm(bad ~ sex + marital_status + age + personal_net_income,
             data = credit, family = binomial(link = logit))
linear_formula(model)
}

