% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds_table.R
\name{odds_table}
\alias{odds_table}
\title{Calculating Odd table}
\usage{
odds_table(score, target, nclass = 10, quantile = TRUE, breaks = NULL)
}
\arguments{
\item{score}{A numeric vector of predictions}

\item{target}{A numeric binary vector {0,1}}

\item{nclass}{A numeric input for determinate the number of classes if \code{breaks} is not given.}

\item{quantile}{A boolean to set if the cuts are made by quantile or counts. This parameter is udes if \code{breaks} is not given.}

\item{breaks}{An optional numeric vector to set the intervals to use}
}
\value{
A data_frame object with the counts, percents and odds
}
\description{
This function calculate a odds table.
}
\examples{
data(predictions)

score <- round(predictions$score * 1000)
target <- predictions$target

odds_table(score, target, nclass = 5)
odds_table(score, target, nclass = 5, quantile = FALSE)
odds_table(score, target, breaks = c(-Inf, 250, 750, Inf))
}

