% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme_rskr.R
\name{palette_rskr}
\alias{palette_rskr}
\title{Riskr color palette}
\usage{
palette_rskr()
}
\description{
Palette based on \emph{Highcharts JS}.
}
\examples{
library("ggplot2")
ggplot(mtcars) +
 geom_point(aes(hp, mpg, color = factor(gear)), size = 4) +
 theme_rskr() +
 scale_color_manual(values = palette_rskr())
}
\references{
\url{https://github.com/highslide-software/highcharts.com/blob/1c33fd375da4f7cf63dba0c3dd74c3f6d0118541/js/parts/Options.js}
}

