% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binning.R
\name{plot.Binning}
\alias{plot.Binning}
\title{Generic Plot method for Binning class}
\usage{
plot.Binning(x, y = NULL, ...)
}
\arguments{
\item{x}{A binning object frome the Binning class}

\item{y}{A parameter}

\item{...}{Other parameters}
}
\description{
Generic Plot method for Binning class
}

