% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/riskr_plot.R
\name{plot.riskr}
\alias{plot.riskr}
\title{Plot a riskr simulation summary}
\usage{
plot.riskr(f, ...)
}
\arguments{
\item{f}{An object of class 'riskr'}

\item{...}{Other plot parameters}
}
\description{
Plot a riskr simulation summary
}

