% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bivariate_analysis.R
\name{plot_ba}
\alias{plot_ba}
\title{Plot Bivariate Analysis}
\usage{
plot_ba(variable, target, labels = TRUE, order.by = NULL)
}
\arguments{
\item{variable}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}

\item{labels}{A par}

\item{order.by}{A par}
}
\description{
This function calculate a bivariate table.
}
\examples{
data("credit")

variable <- credit$sex
target <- credit$bad

plot_ba(variable, target)
plot_ba(variable, target, order.by = "target")
}

