% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bivariate_analysis.R
\name{plot_ba2}
\alias{plot_ba2}
\title{Plot Bivariate Analysis (2)}
\usage{
plot_ba2(variable, target, labels = TRUE, order.by = NULL)
}
\arguments{
\item{variable}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}

\item{labels}{A par}

\item{order.by}{A par}
}
\value{
A ggplot2::ggplot2 object
}
\description{
A minimal version for \emph{plot_ba}
}
\examples{
data("credit")

variable <- as.character(credit$marital_status)
target <- credit$bad

plot_ba2(variable, target)
plot_ba2(variable, target, labels = FALSE)
plot_ba2(variable, target, order.by = "odds")
}

