% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_biv_table.R
\name{plot_biv_table}
\alias{plot_biv_table}
\title{Plot Bivariableiate Analysis}
\usage{
plot_biv_table(variable, target, count.labels = FALSE,
  target.labels = FALSE, arrange.plot.by = "variable", coord.flip = FALSE,
  add.legend = TRUE, legend.color = "gray80", target.color = "darkblue",
  bar.color = "gray80", bar.width = 0.6, rate.size = 1, size.text = 4,
  size.text2 = 10, remove.axis.y = FALSE)
}
\arguments{
\item{variable}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}

\item{count.labels}{A par}

\item{target.labels}{A par}

\item{coord.flip}{A par}

\item{add.legend}{A par}

\item{legend.color}{A par}

\item{target.color}{A par}

\item{bar.width}{A par}

\item{rate.size}{A par}

\item{size.text}{A par}

\item{size.text2}{A par}

\item{remove.axis.y}{A par}
}
\value{
A ggplot2 object
}
\description{
Plot Bivariableiate Analysis
}
\examples{
data("credit")

variable <- credit$quant_add_cards
target <- credit$bad

plot_biv_table(variable, target)
plot_biv_table(variable, target) + theme_light()
plot_biv_table(variable, target) + theme_gray()
plot_biv_table(variable, target, coord.flip = TRUE)
plot_biv_table(variable, target, add.legend = FALSE)
plot_biv_table(variable, target, count.labels = TRUE, coord.flip = TRUE)
plot_biv_table(variable, target, target.labels = TRUE)
plot_biv_table(variable, target, count.labels = TRUE, target.labels = TRUE)
}

