% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bivariate_analysis.R
\name{plot_bt}
\alias{plot_bt}
\title{Bivariate Plot}
\usage{
plot_bt(variable, target, count.labels = FALSE, target.labels = FALSE,
  arrange.plot.by = NULL, coord.flip = FALSE, add.legend = TRUE,
  legend.color = "gray80", target.color = "darkblue",
  bar.color = "gray80", bar.width = 0.6, rate.size = 1, size.text = 4,
  size.text2 = 10, remove.axis.y = FALSE)
}
\arguments{
\item{variable}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}

\item{count.labels}{A par}

\item{target.labels}{A par}

\item{coord.flip}{A par}

\item{add.legend}{A par}

\item{legend.color}{A par}

\item{target.color}{A par}

\item{bar.width}{A par}

\item{rate.size}{A par}

\item{size.text}{A par}

\item{size.text2}{A par}

\item{remove.axis.y}{A par}
}
\value{
A ggplot2 object
}
\description{
Bivariate Plot
}
\examples{
data("credit")

variable <- credit$sex
target <- credit$bad

plot_bt(variable, target)
plot_bt(variable, target) + theme_gray()
plot_bt(variable, target, coord.flip = TRUE)
plot_bt(variable, target, add.legend = FALSE)
plot_bt(variable, target, count.labels = TRUE, coord.flip = TRUE)
plot_bt(variable, target, count.labels = TRUE, target.labels = TRUE)
}

