% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gain}
\alias{plot_gain}
\title{Plot Gain Curve}
\usage{
plot_gain(score, target)
}
\arguments{
\item{score}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}
}
\value{
The KS statistic
}
\description{
Return a ggplot object.
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

plot_gain(score, target)
}

