% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lift}
\alias{plot_lift}
\title{Plot Lift Chart}
\usage{
plot_lift(score, target)
}
\arguments{
\item{score}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}
}
\value{
The plot
}
\description{
Return a ggplot2::ggplot object.
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

plot_lift(score, target)
}
\references{
http://www2.cs.uregina.ca/~dbd/cs831/notes/lift_chart/lift_chart.html http://www.saedsayad.com/model_evaluation_c.htm
}

