% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_perf}
\alias{plot_perf}
\title{Plot Performance}
\usage{
plot_perf(score, target)
}
\arguments{
\item{score}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}
}
\value{
The plot
}
\description{
Return a ggplot2::ggplot object.
}
\examples{
data(predictions)

score <- 1000 * predictions$score
target <- predictions$target

plot_perf(score, target)
}

