% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ranking.R
\name{pred_ranking}
\alias{pred_ranking}
\title{Ranking for predictives variables}
\usage{
pred_ranking(df, target_name = NULL, nuniques = 10)
}
\arguments{
\item{df}{A data frame}

\item{target_name}{The name of response variable}

\item{nuniques}{Limit to consider a numeric varialbe as categorical one}
}
\value{
A data frame with statistics like aucroc and ks from de univariate models `response ~ target`
}
\description{
Calculate the predictive ranking from a data frame.
}
\examples{
data("credit")
df <- credit
target_name <- "bad"

pred_ranking(df, target_name)

}

