% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/riskr.R
\docType{data}
\name{predictions}
\alias{predictions}
\title{simulated dataset: target and predictions}
\format{A data frame with 10000 rows and 2 variables}
\usage{
data(predictions)
}
\description{
A dataset containing the prices and other attributes of almost 54,000
 diamonds. The variables are as follows:
}
\details{
\itemize{
  \item score Numeric variable
  \item target. A binary numeric vector
}
}
\keyword{datasets}

