% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pretty_df.R
\name{pretty_df}
\alias{pretty_df}
\title{Make a data frame looks a bit pretty}
\usage{
pretty_df(df, to.lower.df.names = TRUE, factor.to.string = TRUE,
  replace.num.na.with = NULL, replace.chr.na.with = NULL,
  trim.chr.vars = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{to.lower.df.names}{df' names to lower?}

\item{factor.to.string}{Factor to string?}

\item{replace.num.na.with}{Value to replace NA when there are NAs in a numeric variables}

\item{replace.chr.na.with}{Value to replace NA when there are NAs in a character variables}

\item{trim.chr.vars}{Trim character vars=}
}
\value{
The same df with \code{tbl_df} class, lower names.
}
\description{
This function standarize a data frame converting factors to characters, setting to lowercase the column names.
}
\examples{
df <- data.frame(leTters = LETTERS[1:10], AA = rnorm(10), leTTers2 = letters[1:10])

df[c(1, 4),2] <- NA
df[c(2, 5),1] <- NA

str(df)
str(pretty_df(df))
pretty_df(df, replace.num.na.with = 0)
pretty_df(df, replace.chr.na.with = "no data")
}

