% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binning.R
\name{print.Binning}
\alias{print.Binning}
\title{Generic Print method for Binning class}
\usage{
\method{print}{Binning}(x, ...)
}
\arguments{
\item{x}{A binning object frome the Binning class}

\item{...}{Other parameters}
}
\description{
Generic Print method for Binning class
}

