% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskr.R
\name{riskr-exports}
\alias{\%>\%}
\alias{riskr-exports}
\title{riskr exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the riskr package to avoid listing the magrittr
as Depends of riskr
}

