\name{riskR-package}
\alias{riskR-package}
\docType{package}
\title{
\packageTitle{riskR}
}
\description{
\packageDescription{riskR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{riskR}
}
\author{
\packageAuthor{riskR}

Maintainer: \packageMaintainer{riskR}
}
\references{
Righi, M. (2015). Loss-Deviation risk measures. Working Paper. Available in: http://arxiv.org/abs/1511.06943}