% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/riskr.R
\docType{package}
\name{riskr}
\alias{riskr}
\alias{riskr-package}
\title{A package to calculate Hasbro Risk battle odds.}
\description{
A package to calculate Hasbro Risk battle odds.
}
\author{
Tim CD Lucas
}

