% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/score_indicators.R
\name{score_indicators}
\alias{score_indicators}
\title{Summary of Performance}
\usage{
score_indicators(score, target)
}
\arguments{
\item{score}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}
}
\value{
A 1 row data frame with the summary of the performance.
}
\description{
Summary of Performance
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

score_indicators(score, target)
}

