% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sd_df.R
\name{sd_df}
\alias{sd_df}
\title{Standarize a data frame}
\usage{
sd_df(df, to.lower.df.names = TRUE, factor.to.string = TRUE,
  fill.num.na.with = 0, fill.chr.na.with = "")
}
\arguments{
\item{df}{A data frame}
}
\value{
The same df with \code{tbl_df} class, lower names.
}
\description{
This function standarize a data frame converting factors to characters, setting to lowercase the column names.
}
\examples{
df <- data.frame(leTters = LETTERS[1:10], AA = rnorm(10), leTTers2 = letters[1:10])

df[c(1, 4),2] <- NA
df[c(2, 5),1] <- NA

str(df)
str(sd_df(df))
}

