% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_summ.R
\name{select_categorical}
\alias{select_categorical}
\title{Selecting categorical variables from a data frame}
\usage{
select_categorical(df)
}
\arguments{
\item{df}{A data frame.}
}
\description{
Selecting categorical variables from a data frame
}
\examples{

data("credit")

credit \%>\% select_categorical()

}

