% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_extras.R
\name{str_clean}
\alias{str_clean}
\title{str_clean}
\usage{
str_clean(string, remove.punct = TRUE, remove.digit = TRUE,
  remove.accent = TRUE, trim = TRUE, to.lower = TRUE)
}
\arguments{
\item{string}{A string}
}
\description{
A description.
}

