% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binning.R
\name{summary.Binning}
\alias{summary.Binning}
\title{Generic Summary method for Binning class}
\usage{
\method{summary}{Binning}(object, ...)
}
\arguments{
\item{object}{A binning object frome the Binning class}

\item{...}{Other parameters}
}
\description{
Generic Summary method for Binning class
}

