% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binning.R
\name{sup_bin}
\alias{sup_bin}
\title{Supervised Binning}
\usage{
sup_bin(target, variable)
}
\arguments{
\item{target}{A numeric binary vector {0,1}}

\item{variable}{A variable}
}
\value{
A list of elements
}
\description{
Automatic supervised binning
}
\examples{
data("credit")

target <- credit$bad
variable <- credit$age

sup_bin(credit$bad, credit$age)
}

