% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{superv_bin}
\alias{superv_bin}
\title{Supervised Binning}
\usage{
superv_bin(variable, target, min.p = 0.05, min.cri = 0.95,
  max.depth = Inf)
}
\arguments{
\item{variable}{A variable}

\item{target}{A numeric binary vector {0,1}}

\item{min.p}{Minimal proportion in a group (a ctree_control argument).}

\item{min.cri}{Minimal critetion (a ctree_control argument).}

\item{max.depth}{Maximun depth in the tree}
}
\value{
A list of elements
}
\description{
Automatic supervised binning
}
\examples{

data("credit")

variable <-  credit$age
target <- credit$bad

superv_bin(variable, target)

variable <- credit$marital_status

superv_bin(variable, target)

}

