% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme_rskr.R
\name{theme_rskr}
\alias{theme_rskr}
\title{Riskr ggplot theme}
\usage{
theme_rskr(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Theme based on \emph{Highcharts JS}.
}
\examples{
library("ggplot2")
qplot(hp, mpg, data = mtcars, geom = 'point') + theme_rskr() + ggtitle('Diamond Prices')
}
\references{
\url{http://www.highcharts.com/demo/line-basic}
}

