% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_summ.R
\name{variability}
\alias{variability}
\title{Calculate the variability}
\usage{
variability(x, na.rm = TRUE)
}
\description{
Calculate the variability
}
\examples{

x <- rbinom(100, p = .005, 10)
variability(x)

x <- rnorm(500)
variability(x)

}

