% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/performance.R
\name{aucroc}
\alias{aucroc}
\title{Calculate Area Under ROC Curve}
\usage{
aucroc(target, score)
}
\arguments{
\item{target}{A numeric binary vector (0, 1)}

\item{score}{A numeric vector containing scores or probabilities}
}
\value{
The AUROC Curve
}
\description{
Calculate Area Under ROC Curve
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

aucroc(target, score)
}

