% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bt.R
\name{bt}
\alias{bt}
\title{bivariate Analysis}
\usage{
bt(variable, target)
}
\arguments{
\item{variable}{A variable}

\item{target}{A numeric binary vector {0,1}}
}
\value{
A data_frame object with the counts, percents and odds
}
\description{
This function calculate a bivariate table.
}
\examples{
data(credit)

variable <- credit$marital_status
target <- 1 - credit$bad

bt(variable, target)

variable <- cut(credit$payment_day, breaks = c(-Inf, 10, 20, Inf))

bt(variable, target)
}

