% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conf_matrix.R
\name{conf_matrix}
\alias{conf_matrix}
\title{Calculating A Confusion Matrix}
\usage{
conf_matrix(pred_class, target)
}
\arguments{
\item{pred_class}{A vector numeric binary (0, 1) of predictions}

\item{target}{A numeric binary vector (0, 1)}
}
\value{
A list with: the confusion matrix, some summaries.
}
\description{
Function to calculate the confusion matrix given to vectors
}
\references{
http://www2.cs.uregina.ca/~dbd/cs831/notes/confusion_matrix/confusion_matrix.html
}

