% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskr.R
\docType{data}
\name{credit}
\alias{credit}
\title{credit data}
\format{A data frame with 49694 rows and 17 variables}
\usage{
data(credit)
}
\description{
This dataset classifies people by a set of attributes as good or bad credit risks
}
\keyword{datasets}

