% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/performance.R
\name{divergence}
\alias{divergence}
\title{Calculate Divergence}
\usage{
divergence(score, target)
}
\arguments{
\item{score}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}

\item{percents}{Values to calculate the gain}
}
\value{
The Gini Coefficient
}
\description{
Calculate the divergence between empirical distributions
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

divergence(score, target)
}

