% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_summ.R
\name{ez_summ}
\alias{ez_summ}
\title{Generate Summary Tables for data frames}
\usage{
ez_summ(df, nuniques = 10, target_name = NULL, ...)
}
\arguments{
\item{df}{A data frame.}

\item{nuniques}{Limit to consider a numeric varialbe as categorical one}

\item{target_name}{Target name if bivariate summary is required}

\item{...}{}
}
\value{
A list
}
\description{
Generate Summary Tables for data frames
}
\examples{

data("credit")

credit \%>\% ez_summ()
credit \%>\% ez_summ(target_name = "bad")

}

