% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_summ.R
\name{ez_summ_biv}
\alias{ez_summ_biv}
\title{Generate Summary Tables for numercial variables}
\usage{
ez_summ_biv(df, target_name = NULL, verbose = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{target_name}{The name of response variable}
}
\description{
Generate Summary Tables for numercial variables
}
\examples{

data("credit")

credit \%>\% ez_summ_biv(target_name = "bad") 


}

