% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ft.R
\name{ft}
\alias{ft}
\title{Create a frecueny table from a vector}
\usage{
ft(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The frecuency table of \code{x}. Frecuencies and percents are returned in a
\code{data.frame} object (adding \code{tbl_df} class).
}
\description{
Create a frecueny table from a vector
}
\examples{
set.seed(1313)
x <- sample(letters[1:10], size = 100, prob = 1:10/10, replace = TRUE)
ft(x)
}

