% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{gg_cum}
\alias{gg_cum}
\title{Plot to compare cumulatives distrbutions}
\usage{
gg_cum(target, score)
}
\arguments{
\item{target}{A numeric binary vector (0, 1)}

\item{score}{A numeric vector containing scores or probabilities}
}
\value{
The plot
}
\description{
Return a plot with ecdfs
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

gg_cum(target, score)
}

