% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{gg_gain}
\alias{gg_gain}
\title{Plot Gain Curve}
\usage{
gg_gain(target, score)
}
\arguments{
\item{target}{A numeric binary vector (0, 1)}

\item{score}{A numeric vector containing scores or probabilities}
}
\value{
the plot
}
\description{
Return a plot for gains
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

gg_gain(target, score)
}

