% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plots_perf.R
\name{gg_roc}
\alias{gg_roc}
\title{Plot AUC Curve}
\usage{
gg_roc(target, ...)
}
\arguments{
\item{target}{A numeric binary vector (0, 1)}

\item{...}{A data frame of scores, or a multiples score vectors: score1, score2, etc}
}
\value{
the plot
}
\description{
Plot AUC Curve
}
\examples{

set.seed(1313)
n <- 5000
score <- runif(n)
score_2 <- score + runif(n)
score_3 <- score + runif(n)
target <- rbinom(n, 1, prob = score)

df <- data.frame(target, score, score_2, score_3)

gg_roc(target, score)
gg_roc(target, model = score, score_3)

library("dplyr")

gg_roc(target,  df \%>\% select(starts_with("score")))

}

