% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{ks}
\alias{ks}
\title{Calculate Kolmogorov-Smirnov statistic}
\usage{
ks(target, score)
}
\arguments{
\item{target}{A numeric binary vector (0, 1)}

\item{score}{A numeric vector containing scores or probabilities}
}
\value{
The KS statistic
}
\description{
Calculate Kolmogorov-Smirnov statistic
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

ks(target, score)

}

