% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{perf}
\alias{perf}
\title{Summary of Performance}
\usage{
perf(target, ...)
}
\arguments{
\item{target}{A numeric binary vector (0, 1)}

\item{...}{A data frame of scores, or a multiples score vectors: score1, score2, etc}
}
\value{
A 1 row data frame with the summary of the performance.
}
\description{
Summary of Performance
}
\examples{
data(predictions)

score <- predictions$score
target <- predictions$target

perf(target, score)
perf(target, score, score2 = score + runif(length(score)))
}

