% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_biv_table.R
\name{plot_biv_table}
\alias{plot_biv_table}
\title{Plot Bivariate Analysis}
\usage{
plot_biv_table(var, target, count.labels = FALSE, target.labels = FALSE,
  coord.flip = FALSE, add.legend = TRUE, legend.color = "gray80",
  rate.color = "darkred", bar.color = "gray80", bar.width = 0.6,
  rate.size = 1, size.text = 4, size.text2 = 10, remove.axis.y = FALSE)
}
\arguments{
\item{var}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}

\item{count.labels}{A par}

\item{target.labels}{A par}

\item{coord.flip}{A par}

\item{add.legend}{A par}

\item{legend.color}{A par}

\item{rate.color}{A par}

\item{bar.width}{A par}

\item{rate.size}{A par}

\item{size.text}{A par}

\item{size.text2}{A par}

\item{remove.axis.y}{A par}
}
\value{
A ggplot2 object
}
\description{
Plot Bivariate Analysis
}
\examples{
data("credit")

var <- credit$sex
target <- credit$bad

plot_biv_table(var, target)
plot_biv_table(var, target) + theme_light()
}

