% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bivariate_analysis.R
\name{plot_bt}
\alias{plot_bt}
\title{Bivariate Plot}
\usage{
plot_bt(variable, target, labels = FALSE, arrange.plot.by = NULL)
}
\arguments{
\item{variable}{A numeric vector containing scores or probabilities}

\item{target}{A numeric binary vector (0, 1)}

\item{labels}{A par}

\item{arrange.plot.by}{A par}
}
\value{
A ggplot2 object
}
\description{
Bivariate Plot
}
\examples{
data("credit")

variable <- credit$sex
target <- credit$bad

plot_bt(variable, target)
plot_bt(variable, target) + theme_gray()
plot_bt(variable, target, labels = TRUE)
}

