% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pred_ranking.R
\name{pred_ranking}
\alias{pred_ranking}
\title{Ranking for predictives variables}
\usage{
pred_ranking(df, target_name = "target", verbose = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{response}{The name of response variable}
}
\value{
A data frame with statistics like aucroc and ks from de univariate models `response ~ target`
}
\description{
Calculate the predictive ranking from a data frame.
}
\examples{
data("credit")
df <- credit
target_name <- "bad"
}

