% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{sd_df}
\alias{sd_df}
\title{Standarize a data frame}
\usage{
sd_df(df, to.lower.df.names = TRUE, factor.to.string = TRUE,
  fill.num.na = TRUE, fill.chr.na = TRUE)
}
\arguments{
\item{df}{A data frame}
}
\value{
The same df with \code{tbl_df} class, lower names.
}
\description{
This function standarize a data frame converting factors to characters, setting to lowercase the column names.
}
\examples{
df <- data.frame(LETTERS, AA = rnorm(length(LETTERS)), leTTers2 = letters)
df[c(1, 4),2] <- NA
str(df)
str(sd_df(df))
}

