% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_summ.R
\name{select_numeric}
\alias{select_numeric}
\title{Selecting numeric variables from a data frame}
\usage{
select_numeric(df, nuniques = 10)
}
\arguments{
\item{df}{A data frame.}

\item{nuniques}{Limit to consider a numeric varialbe as categorical one}
}
\description{
Selecting numeric variables from a data frame
}
\examples{

data("credit")

credit \%>\% select_numeric()

}

