% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_summ.R
\name{unvariability}
\alias{unvariability}
\title{Calculate the unvariability}
\usage{
unvariability(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{A logical indicating whether missing values should be removed.}
}
\description{
Calculate the unvariability
}
\examples{

x <- rbinom(100, p = .008, 10)
unvariability(x)

x <- rnorm(500)
unvariability(x)

}

